Criação de validações e endpoints

+++ src/controles/ControleCategoria.js
@@ -43,6 +43,12 @@ const ControleCategoria = {
     const { id } = req.params;
 
     try {
+      const validaId = await db.query(`SELECT * FROM categoria WHERE id = $1`, [
+        id,
+      ]);
+      if (validaId.rows.length === 0) {
+        return res.status(404).json({ error: "Categoria nao encontrada" });
+      }
       const resultado = await db.query(
         "DELETE FROM categoria WHERE id = $1 RETURNING *",
         [id]
@@ -50,13 +56,42 @@ const ControleCategoria = {
 
       if (resultado.rowCount > 0) {
         res.status(204).json({});
+      } else {
+        res.status(304).json({});
       }
-
-      res.status(304).json({});
     } catch (error) {
       res.status(500).json({ error: error.message });
     }
   },
-  async update(req, res) {},
+  async update(req, res) {
+    const { id } = req.params;
+    const { nome, descricao } = req.body;
+
+    try {
+      const validaId = await db.query(`SELECT * FROM categoria WHERE id = $1`, [
+        id,
+      ]);
+      if (validaId.rows.length === 0) {
+        return res.status(404).json({ error: "Categoria nao encontrada" });
+      }
+
+      const updateCategoria = await db.query(
+        ` UPDATE categoria 
+            SET  nome = $1, descricao = $2
+            WHERE id = $3
+            RETURNING *; `,
+
+        [nome, descricao, id]
+      );
+
+      if (updateCategoria.rowCount > 0) {
+        res.status(200).json(updateCategoria.rows[0]);
+      } else {
+        res.status(304).json({ error: "atualização nao encontrada" });
+      }
+    } catch (error) {
+      res.status(500).json({ error: error.message });
+    }
+  },
 };
 module.exports = ControleCategoria;
